:ram: Split results page tests from other app tests
